test(snapshot-update): Merge migrate test into update test
Consolidate snapshot_migrate_test.py content into snapshot_update_test.py. Update all test invocations to use file paths instead of UUIDs. Add TestWithLifecycleHooks and TestWithLifecycleHooksAndScenario test classes. Remove snapshot_migrate_test.py as it is no longer needed.
